[WIP] First version with CAEN converter working
Still a lot of print-outs must be cleaned out.
Missing the proper conversion of the channels to col-row hits.
Also missing how to show the waveform, and how to extract the
value to fill the charge in the hit (so far is the amplitude
of the waveform).
There is a check in the raw data (size==0) which I'm not sure if
can be there, must be checked.