[WIP] Include new timing hit in the online monitor
Timing pixels are identified with a new histo class. The needed
infrastructure to monitor them is being built.
The DAQ assumed is a digitizer (see caen-dt5742 module), where a
sensor could be read by different channels from different
digitizers. That implies the necessity to create a new hit class
(TimingHitHistos) to deal with this, and assing as plane unit the
sensor (dut) which could be in different digitizers, and therefore
provided by different producers.